Updates models to return 404 errors in relevant cases

+++ src/models/plot-models.ts
@@ -2,11 +2,13 @@ import QueryString from "qs"
 import { db } from "../db"
 import format from "pg-format"
 import { Plot } from "../types/plot-types"
-import { checkPlotNameConflict, getPlotOwnerId, validatePlotType } from "../utils/db-query-utils"
-import { verifyPermission, verifyResult } from "../utils/verification-utils"
+import { checkPlotNameConflict, getPlotOwnerId, searchForUserId, validatePlotType } from "../utils/db-query-utils"
+import { verifyPermission } from "../utils/verification-utils"
 
 
 export const selectPlotsByOwner = async (authUserId: number, owner_id: number, { type }: QueryString.ParsedQs): Promise<Plot[]> => {
+
+  await searchForUserId(owner_id)
 
   await verifyPermission(authUserId, owner_id, "Permission to view plot data denied")
 
@@ -35,6 +37,8 @@ export const selectPlotsByOwner = async (authUserId: number, owner_id: number, {
 
 export const insertPlotByOwner = async (authUserId: number, owner_id: number, plot: Plot): Promise<Plot> => {
 
+  await searchForUserId(owner_id)
+
   await verifyPermission(authUserId, owner_id, "Permission to create plot denied")
 
   await checkPlotNameConflict(owner_id, plot.name)
@@ -58,13 +62,19 @@ export const insertPlotByOwner = async (authUserId: number, owner_id: number, pl
     `,
     [plot.owner_id, plot.name, plot.type, plot.description, plot.location, plot.area])
 
-  await verifyResult(!result.rowCount, "Plot not found")
-
   return result.rows[0]
 }
 
 
 export const selectPlotByPlotId = async (authUserId: number, plot_id: number): Promise<Plot> => {
+
+  if (isNaN(plot_id)) {
+    return Promise.reject({
+      status: 404,
+      message: "Not Found",
+      details: "Plot not found"
+    })
+  }
 
   const owner_id = await getPlotOwnerId(plot_id)
 
@@ -77,13 +87,19 @@ export const selectPlotByPlotId = async (authUserId: number, plot_id: number): P
     `,
     [plot_id, owner_id])
 
-  await verifyResult(!result.rowCount, "Plot not found")
-
   return result.rows[0]
 }
 
 
 export const updatePlotByPlotId = async (authUserId: number, plot_id: number, plot: Plot): Promise<Plot> => {
+
+  if (isNaN(plot_id)) {
+    return Promise.reject({
+      status: 404,
+      message: "Not Found",
+      details: "Plot not found"
+    })
+  }
 
   const owner_id = await getPlotOwnerId(plot_id)
 
@@ -123,24 +139,28 @@ export const updatePlotByPlotId = async (authUserId: number, plot_id: number, pl
     `,
     [plot.name, plot.type, plot.description, plot.location, plot.area, plot_id])
 
-  await verifyResult(!result.rowCount, "Plot not found")
-
   return result.rows[0]
 }
 
 
 export const removePlotByPlotId = async (authUserId: number, plot_id: number): Promise<void> => {
 
+  if (isNaN(plot_id)) {
+    return Promise.reject({
+      status: 404,
+      message: "Not Found",
+      details: "Plot not found"
+    })
+  }
+
   const owner_id = await getPlotOwnerId(plot_id)
 
   await verifyPermission(authUserId, owner_id, "Permission to delete plot data denied")
 
-  const result = await db.query(`
+  await db.query(`
     DELETE FROM plots
     WHERE plot_id = $1
     RETURNING *;
     `,
     [plot_id])
-
-  await verifyResult(!result.rowCount, "Plot not found")
 }
